Better display of operations.
Operations.to_string has been rewritten to take into account the
previous equation. This way, multiplications and divisions are displayed
only for terms that are present in the previous Equation.
Also, it has improved the display of an equation where at least one
member is made of only one term.